I added type parameters to Vars and LambdaExpressions in substitutions. Since the "level" of a formula is irrelevant for a substitution, I added dummy type parameters: e.g. Var[_] and LambdaEpression[_] ...
But then the compiler complains about wrong inference of type parameters.

I think I could solve this problem by parametering the Substitution classes by a type T (lowerbound by Lambda) and replacing the dummy type "_" by this type T everywhere...   But I find this (i.e. the necessity to have the "layer" as a type parameter in every algorithm that we do) not very elegant...

@Tomer: Do you know any way to solve these type parameter problems without parameterizing the Substitution classes ??

git-svn-id: http://gapt.googlecode.com/svn/trunk/source@120 85d21ef8-add0-11de-ae77-5bfb9c5a3f34