Draft for an alternative approach to language layers and lambda calculus.
This draft is based on the following:
The idea is to have factories, but to change how they are accessed. Instead of using the type inference mechanism, in the attached draft the factory is determined by a global singleton ExpressionFactory. This is of course not ideal: If one wants to create expressions from two languages such that neither is contained in the other (e.g. propositional linear logic and classical higher order logic), one has to switch out the factories. I think this is not the case for languages such that one contains the other, there one could simply use the factory of the more narrow language.

I will try to think of alternative ways to control which factory is used, except singletons and type parameters.



git-svn-id: http://gapt.googlecode.com/svn/trunk/source@142 85d21ef8-add0-11de-ae77-5bfb9c5a3f34