Hi All,
for the unification algorithm I should apply unification in the substitution.scala. This is neede for the case : {y->f1, y->f2}. The terms f1 and f2 should unify. But when I import at.logic.unification.FOLUnification._   in substitution.scala, an error occurs. The error message is :  value unification is not a member of package at.logic

The function applySubst is commented. It applies the substitution sub to this.


git-svn-id: http://gapt.googlecode.com/svn/trunk/source@249 85d21ef8-add0-11de-ae77-5bfb9c5a3f34