Proof Profile looks good now.
I incorporated Cvetan's changes to the Struct (Times now records the auxiliary formulas of the inference from which it comes) into Struct.scala, and changed the standard clause set accordingly. 

The proof profile is now computed by first applying Bruno's rewrite rules, and the normalizing the struct and transforming it into a set of clauses.



git-svn-id: http://gapt.googlecode.com/svn/trunk@537 85d21ef8-add0-11de-ae77-5bfb9c5a3f34