Correcting the size of the smallest grammar
On the method for cut-introduction using maxsat, a
list of grammars is returned because of the
current implementation that supports only grammars
of the type U o S. Since the method needs more
general grammars, i.e., U o S1 o ... o Sn, this is
achieved by having a list of simpler grammars.
Consequently the computation of the size of these
grammars needs some adaptation.

git-svn-id: http://gapt.googlecode.com/svn/trunk@2338 85d21ef8-add0-11de-ae77-5bfb9c5a3f34