Using lists for clause sets
The implementation of Set in scala is less
efficient than that of List. As a result, the
Tseitin transformation was taking 90% of the time
for the cut-introduction method using maxsat.
Changing the type only inside Tseiting brought the
time down, but the final transformation of a list
into a set was still taking 70% of the total time.
In this commit this change is propagated to the
interface and therefore to all other uses of
clause sets in gapt. With this change, the Tseitin
transformation now takes 1 to 2% of the time in
the cut-introduction algorithm.

git-svn-id: http://gapt.googlecode.com/svn/trunk@2433 85d21ef8-add0-11de-ae77-5bfb9c5a3f34